requirements.txt
fastapi
uvicorn
ultralytics
mlflow
prefect
Pillow
pytest
